service/requirements.txt
Flask==1.1.4
flask_caching==1.8.0
python-dotenv==0.18.0
redis==3.5.3
